refactor: 공용 스타일 컴포넌트 구조 정리 및 일부 수정
