+++ src/components/shared/TextStyles.ts
@@ -3,11 +3,9 @@ import styled from '@emotion/styled';
 export const Title = styled.p`
   font-size: 20px;
   font-weight: 600;
-  text-align: center;
 `;
 export const SubTitle = styled.p`
   font-size: 14px;
   font-weight: 500;
   color: #7c7c7c;
-  text-align: center;
 `;
