+++ src/components/shared/UIStyles.ts
@@ -1,10 +1,13 @@
 import styled from '@emotion/styled';
+// 공용으로 쓰이는 UI 스타일 컴포넌트
 export const Container = styled.div`
   display: flex;
   flex-direction: column;
+  align-items: center;
 `;
 export const Header = styled.div`
   display: flex;
   flex-direction: column;
+  align-items: center;
   gap: 7px;
 `;
