+++ src/pages/chatRoomExit/chatRoomSummaryPage.tsx
@@ -1,7 +1,7 @@
 import styled from '@emotion/styled';
 import { useState } from 'react';
 import Button from '@components/chatRoomExit/Button';
-import { Container, Header } from '@components/chatRoomExit/UIStyles';
+import { Container, Header } from '@components/shared/UIStyles';
 import { Title, SubTitle } from '@components/shared/TextStyles';
 import SummaryModal from '@components/chatRoomExit/SummaryModal';
 import { ModalPortal } from '@components/shared/ModalPortal';
